// sdk/src/models/errors/structured-error.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v3";
import * as types from "../../types/primitives.js";
import { PumbleSDKError } from "./pumble-sdk-error.js";

/**
 * Structured validation error from the framework layer.
 */
export type StructuredErrorData = {
  /**
   * Raw error message.
   */
  message: string;
  /**
   * Localized variant of `message` (often identical for English clients).
   */
  localizedMessage: string;
  /**
   * Internal error code.
   */
  code: number;
};

/**
 * Structured validation error from the framework layer.
 */
export class StructuredError extends PumbleSDKError {
  /**
   * Localized variant of `message` (often identical for English clients).
   */
  localizedMessage: string;
  /**
   * Internal error code.
   */
  code: number;

  /** The original data that was passed to this error instance. */
  data$: StructuredErrorData;

  constructor(
    err: StructuredErrorData,
    httpMeta: { response: Response; request: Request; body: string },
  ) {
    const message = err.message || `API error occurred: ${JSON.stringify(err)}`;
    super(message, httpMeta);
    this.data$ = err;
    this.localizedMessage = err.localizedMessage;
    this.code = err.code;

    this.name = "StructuredError";
  }
}

/** @internal */
export const StructuredError$inboundSchema: z.ZodType<
  StructuredError,
  z.ZodTypeDef,
  unknown
> = z.object({
  message: types.string(),
  localizedMessage: types.string(),
  code: types.number(),
  request$: z.instanceof(Request),
  response$: z.instanceof(Response),
  body$: z.string(),
})
  .transform((v) => {
    return new StructuredError(v, {
      request: v.request$,
      response: v.response$,
      body: v.body$,
    });
  });

/** @internal */
export type StructuredError$Outbound = {
  message: string;
  localizedMessage: string;
  code: number;
};

/** @internal */
export const StructuredError$outboundSchema: z.ZodType<
  StructuredError$Outbound,
  z.ZodTypeDef,
  StructuredError
> = z.instanceof(StructuredError)
  .transform(v => v.data$)
  .pipe(z.object({
    message: z.string(),
    localizedMessage: z.string(),
    code: z.number().int(),
  }));